backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv # For local development, not used in Next.js
